requirements.txt
fitz
PyMuPDF
pandas
tqdm
google-genai
python-dotenv
pypdf
pyarrow
numpy
neo4j
faiss-cpu
qdrant-client
pytest
